Migrate a bunch of code from public header files into cpp files
