+++ cplusplus/unconvertDWfile.cpp
@@ -16,9 +16,10 @@
 
 #include "zdw/UnconvertFromZDW.h"
 
+#include <algorithm>
 #include <stdio.h>
 #include <stdlib.h>
-#include <algorithm>
+#include <string.h>
 
 using namespace adobe::zdw;
 using std::string;
